Refactor: Update import paths for PropertyManager and catalog_types; add PropertyManager class to manage entry properties
