+++ GeoBasis_Loader_main.py
@@ -4,14 +4,13 @@ from typing import Union, Optional
 from qgis.PyQt.QtWidgets import QMenu, QAction
 from qgis.PyQt.QtGui import QIcon, QColor, QDesktopServices
 from qgis.PyQt.QtCore import QUrl, QObject
-from qgis.core import QgsSettings, QgsProject, QgsVectorLayer, QgsRasterLayer, QgsVectorTileLayer, QgsLayerTree, QgsSymbolLayer, QgsWkbTypes, Qgis
+from qgis.core import QgsSettings, QgsProject, QgsVectorLayer, QgsRasterLayer, QgsVectorTileLayer, QgsSymbolLayer, QgsWkbTypes, Qgis
 from qgis.gui import QgisInterface
 from .topic_search import SearchFilter
 from . import config
 from . import custom_logger
 from . import ui as custom_ui
 from .catalog_manager import CatalogManager
-from .property_manager import singleton as PropertyManager
 from .topic_handlers import catalog_types
 
 logger = custom_logger.get_logger(__file__)
